html/semantics/embedded-content/the-video-element/video-loading-lazy-in-viewport.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  let windowLoadFired = false;

  async_test(t => {
    const video = document.createElement("video");
    video.id = "target";
    video.src = "/media/A4.mp4?pipe=trickle(d2)";
    video.loading = "lazy";

    window.addEventListener("load", () => {
      windowLoadFired = true;
    });

    video.addEventListener("loadstart", () => {
      assert_true(windowLoadFired, "Window load event should fire before loadstart for lazy video in viewport");
      t.done();
    });

    document.body.appendChild(video);
  }, "In-viewport video with loading='lazy' fires window load before loadstart");
</script>
